client/requirements.txt
PyInstaller==3.2.1
requests==2.10.0
pyserial==3.3